Implement basic Vivi customisations
This is basically a straight port from our existing implementation, but
with everything shoved into `BOOTCMD` because the `ENV` constant didn't
seem to work. We also fall back to just booting from partition 1 if
partition 3 can't be read, to handle self-expansion cases.

The `-mno-unaligned-access` option is added to `CFLAGS` to prevent some
CPU abort with displaying the splash screen bitmap.

SolidRun logo is now replaced with a Vivi logo, since that screen is now
needed so we can debug over serial in future.